refactoring

+++ src/services/auth.service.ts
@@ -1,6 +1,6 @@
 import { Request } from "express";
-import jwt, { TokenExpiredError } from 'jsonwebtoken';
-import { checkUserTokenExist, deleteUserToken } from "./Auth/token.service";
+import jwt, { TokenExpiredError } from "jsonwebtoken";
+import { checkUserTokenExist, deleteUserToken } from "./auth/token.service";
 
 function getToken(req: Request) {
     let token: string;
@@ -17,20 +17,22 @@ function getToken(req: Request) {
 async function verifyUserToken(userToken: string) {
     let userId: number | undefined = undefined;
     try {
-        userId = (jwt.verify(userToken, process.env.JWT_SECRET as string, { ignoreExpiration: true }) as Record<string, any>).id;
-        if (!userId)
-            throw new Error()
+        userId = (
+            jwt.verify(userToken, process.env.JWT_SECRET as string, {
+                ignoreExpiration: true,
+            }) as Record<string, any>
+        ).id;
+        if (!userId) throw new Error();
         await checkUserTokenExist(userId, userToken);
         // check expiration of token
         jwt.verify(userToken, process.env.JWT_SECRET as string);
         return userId;
     } catch (err: any) {
         if (err instanceof TokenExpiredError) {
-            console.log('expired');
-            if (userId)
-                deleteUserToken(userId, userToken);
+            console.log("expired");
+            if (userId) deleteUserToken(userId, userToken);
         }
-        throw new Error('Login again.');
+        throw new Error("Login again.");
     }
 }
